supervisor默认设置

+++ supervisor/installSupervisor.sh
@@ -3,7 +3,9 @@ yum install -y epel-release
 yum install -y supervisor
 
 # 备份配置文件，新增个性化配置（无配置则使用默认配置）
+if [ ! -f "/etc/supervisord.example.conf" ];then
 mv -fu /etc/supervisord.conf /etc/supervisord.example.conf
+fi
 
 cat <<EOF> /etc/supervisord.conf
 [unix_http_server]
